A* Implémenté et opérationnel ! Mini 70 FPS
Script A* implémenté
Modifications mineures dans WorldState (int remplacé par float)
Vecteur de direction envoyé par "BulletPath" corrigé

+++ Assets/Scripts/WorldState.cs
@@ -32,7 +32,7 @@ public class WorldState : MonoBehaviour
         }*/
 	}
 
-    public void updateEntitie(Vector3 position,Vector3 ADirection,Quaternion Rotation, int EID,bool isActive,int speed)
+    public void updateEntitie(Vector3 position,Vector3 ADirection,Quaternion Rotation, int EID,bool isActive,float speed)
     {
         for (int i = 0; i < entitiesStruct.Length; i++)
         {
